Now works with JaLoader (before, it was only for BepInEx)
