--- CameraPlacements.csproj
@@ -1,56 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net35</TargetFramework>
-    <AssemblyName>CameraPlacements</AssemblyName>
-    <Description>My first plugin</Description>
-    <Version>1.0.0</Version>
-    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-    <LangVersion>latest</LangVersion>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Theraot.Core" Version="3.2.11" />
-    <PackageReference Include="UnityEngine.Modules" Version="2017.4.23" IncludeAssets="compile" />
-
-	<Reference Include="Assembly-CSharp"> 
-	<HintPath> lib\Assembly-CSharp.dll </HintPath> 
-	</Reference>
-	<Reference Include="Assembly-CSharp-firstpass"> 
-	<HintPath> lib\Assembly-CSharp-firstpass.dll </HintPath> 
-	</Reference>
-	<Reference Include="Assembly-UnityScript"> 
-	<HintPath> lib\Assembly-UnityScript.dll </HintPath> 
-	</Reference>
-	<Reference Include="Assembly-UnityScript-firstpass"> 
-	<HintPath> lib\Assembly-UnityScript-firstpass.dll </HintPath> 
-	</Reference>
-
-	<Reference Include="JaLoader">
-		<HintPath>lib\JaLoader.dll</HintPath>
-	</Reference>
-
-	<Reference Include="UnityEngine">
-		<HintPath>lib\UnityEngine.dll</HintPath>
-	</Reference>
-
-	<Reference Include="UnityEngine.CoreModule">
-	  <HintPath>lib\UnityEngine.CoreModule.dll</HintPath>
-	</Reference>
-  </ItemGroup>
-  
-  <ItemGroup Condition="'$(TargetFramework.TrimEnd(`0123456789`))' == 'net'">
-    <PackageReference Include="Microsoft.NETFramework.ReferenceAssemblies" Version="1.0.2" PrivateAssets="all" />
-  </ItemGroup>
-  
-  <ItemGroup>
-    <Folder Include="Documentation\" />
-  </ItemGroup>
-    <PropertyGroup>
-      <Username Condition="'$(Username)' == ''">$([System.Environment]::UserName)</Username>
-    </PropertyGroup>
-
-    <Target Name="CopyDllToModFolder" AfterTargets="Build">
-      <Copy SourceFiles="$(TargetPath)" DestinationFolder="C:\Users\$(Username)\Documents\Jalopy\Mods" />
-    </Target>
-</Project>
