--- CameraPlacementsAssets.manifest
@@ -1,28 +1,0 @@
-ManifestFileVersion: 0
-CRC: 3531174878
-Hashes:
-  AssetFileHash:
-    serializedVersion: 2
-    Hash: c493d12b21eaca29c1db5fba08e6d7e9
-  TypeTreeHash:
-    serializedVersion: 2
-    Hash: 6fa9ce387392089bc59dd886979cd053
-HashAppended: 0
-ClassTypes:
-- Class: 1
-  Script: {instanceID: 0}
-- Class: 4
-  Script: {instanceID: 0}
-- Class: 21
-  Script: {instanceID: 0}
-- Class: 23
-  Script: {instanceID: 0}
-- Class: 33
-  Script: {instanceID: 0}
-- Class: 43
-  Script: {instanceID: 0}
-- Class: 48
-  Script: {instanceID: 0}
-Assets:
-- Assets/Assets/AssetBundles/pointer-cone.prefab
-Dependencies: []
